#278 NoNulls for Text
